+++ src/test/java/com/braintreegateway/integrationtest/PaymentMethodNonceIT.java
@@ -235,10 +235,11 @@ public class PaymentMethodNonceIT extends IntegrationTest {
         assertTrue(foundNonce.getThreeDSecureInfo().isLiabilityShifted());
         assertTrue(foundNonce.getThreeDSecureInfo().isLiabilityShiftPossible());
         assertNotNull(foundNonce.getThreeDSecureInfo().getStatus());
-        assertNotNull(foundNonce.getThreeDSecureInfo().getCAVV());
-        assertNotNull(foundNonce.getThreeDSecureInfo().getXID());
-        assertNotNull(foundNonce.getThreeDSecureInfo().getECIFlag());
-        assertNotNull(foundNonce.getThreeDSecureInfo().getThreeDSecureVersion());
+        assertEquals("test_cavv", foundNonce.getThreeDSecureInfo().getCAVV());
+        assertEquals("test_xid", foundNonce.getThreeDSecureInfo().getXID());
+        assertEquals("test_eci", foundNonce.getThreeDSecureInfo().getECIFlag());
+        assertEquals("1.0.2", foundNonce.getThreeDSecureInfo().getThreeDSecureVersion());
+        assertEquals((String)null, foundNonce.getThreeDSecureInfo().getDsTransactionId());
     }
 
     @Test